+++ zimui/src/components/channel/tabs/LivesTab.vue
@@ -2,15 +2,15 @@
 import { useMainStore } from '@/stores/main'
 import { computed } from 'vue'
 
-import TabView from '@/views/TabView.vue'
+import LivesGridTab from './LivesGridTab.vue'
 
 const main = useMainStore()
-const livesAvailable = computed(() => main.channel?.livesPlaylist)
+const livesAvailable = computed(() => main.channel?.userLivesPlaylist)
 </script>
 
 <template>
   <div v-if="livesAvailable">
-    <tab-view :playlist-label="'livesPlaylist'" :playlist-type="'Lives'" />
+    <lives-grid-tab/>
   </div>
   <div v-else>
     <p style="text-align: center">No Live available</p>