+++ zimui/src/components/channel/tabs/ShortsTab.vue
@@ -2,18 +2,16 @@
 import { useMainStore } from '@/stores/main'
 import { computed } from 'vue'
 
-import TabView from '@/views/TabView.vue'
+import ShortsGridTab from './ShortsGridTab.vue'
+
 
 const main = useMainStore()
-const shortsAvailable = computed(() => main.channel?.shortsPlaylist)
+const shortsAvailable = computed(() => main.channel?.userShortUploadsPlaylist)
 </script>
 
 <template>
   <div v-if="shortsAvailable">
-    <tab-view 
-    :playlist-label="'shortsPlaylist'" 
-    :playlist-type="'Shorts'"
-    />
+    <shorts-grid-tab/>
   </div>
   <div v-else>
     <p style="text-align: center">No short available</p>