+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Diffrentiate shorts, lives, & normal videos (#367)
+- Diffrentiate user uploaded shorts, lives, & long videos (#367)
 - corrected the short video resolution in the UI (#366)
 
 ### Changed